Added Class Color Picker (see description)
Class Color Picker is for cases where an developer will want to return a class value rather than the hexcode.

This could be used in cases where the color is simply a representation of the idea of what that class looks like.

Basicially it will return "green" rather than "#00ff00" as its value. As such the label is the hexcode color and the value is the class string.

This property editor only supports key value pair prevalues and as such this will increase the minimum Umbraco version required to v7.5.4.